Make changes to the theming of application
- Replace `themeHandeler` function with new hook called `useCustomTheme`
and this hook contain the same theme of app with some additional settings
and code enhancement.

- Delete `ThemeContext` and `useModeType` hook whome are responsible for providing
theme mode type and toggling theme function in the app, And replace them with `useCustomTheme`
which is does the same functionality using `useCookies` hook from `react-cookie` library.

- Convert `html` and `body` elements to MUI `Box` component and add some styles to `html` element.

- Add `renderApp` state to `useUserLogging` hook with 'false' as initial value
and it will stted to 'true' after the app runed for the first time and then `useEffict` hook runs;
this state will used in `AppWrapper` component for prevent the first render of
the main content of app by adding a condition if `renderApp` is stted to 'true' then render content of app.

- Remove an unused class in `global.css` file.

- Update "@abdulrhmangoni/am-store-library" library.

- Install `normalize.css` for reset the default styles of the user agent.

- Install "react-input-emoji" package for a future use.