Enhance email verification page
- Make verification field thows an error as tooltip to tell the user if he enters
invalid input.
Add two image to use them as illustraion images.

- Handle two error cases, "code expired" and "too many requests";
an error message will appears with an illustration image to descripe the case
based on the response of the request of `submitCode` function, if returns
an error with the status code `408`, that is means "time out" ("code expired" case),
And if the status code is `429`,  that is means the case "too many requests".

- Add two illustration images to `/public` folder for displaying in "time out" case
and "too many requests" case.