Enhance and Refactor favorites page
- Create `useUserFavorites` hook and separate handling data logic from
the UI of favorites page to this new hook.

- Cache the loaded favorite products in browser's session storage
preventing the same products to load when every navigation to
favorites page.

- Create `getFavoritesFromSession`, `toggleFavoritesFromSession`,
`clearFavoritesSession` and `setFavoritesSession` functions that handle
the dealing with session storage in `useFavoritesActions` hook.

- Refactor `ToggleFavorite` component for making it adds the
products to favorites slice and to browser's session storage
through `toggleFavoritesFromSession` function, therefore
the component will receives a whole product as prop
instead of product id.

- Use the stored favorite products in browser's session storage
in `CustomFavoriteIcon` component to display the count of
products in user's favorites in case the user is a guest and he
refreshed the page, so that his favorite products won't be lost
unless he closes the app.

- Clear the session storage that caches user's favorite products
through `logout` function in `AccountMenu` component when
the user wants to log out.