Refactor search for products page
- Delete `app/products/error.js` file and make `app/products/page.js`
which is the root or the parent component of search for products page
catches the errors using try-catch blocks because it a server component
and it can be an asynchronous function.

- Delete `SearchResultsPage` component that was fetches all search
result in one request and make `app/products/page.js` replaces it,
But make `app/products/page.js` only fetches the first slice or part
of search results on the server ( because it a server component )
and passes the results to the client component (its child component)
to displays them handles bringing the rest of search results.

- Create a client component (`SearchResultsDisplayer`) for displaying
search results, Make it takes the first slice or part of search results from
its parent (`app/products/page.js` that has feched it from the server),
And brings the rest of search results using infinit scroll logic.

- Rename `currentUrl` variable in `useProductsSearchParams` hook
to `currentPath` and use `replace` method from `useRouter` hook
instead of `push` method to update site's url for avoiding an
unnecessary filling of browsing history.

- Use options object of `useHTTPRequestState` hook to set a default
value to `isError` state in `BestProductsSlideUI` component instead of
using `useEffect` hook to update the state when component rendred.

- Move `NotFound` component from the deleted file `app/products/error.js`
into a new file named `app/products/NotFound.js`

- Update `@abdulrhmangoni/am-store-library`