Refactor payment methods management componets
- Separate managment logic from `PaymentMethodsManagement` componet
into a new separated hook named `usePaymentMethodsManagement`.

- Move the section of "Cards list" button and "Add card" button into
a separated component named `OptionsBar`.

- Move "Choosed Card" title from `PaymentMethodsManagement` componet
into `SelectedCridetCard` componet.

- Remove `selectedCardNumber` prop from `CreditCard`
component and add to it another prop called `isSelectedCard`
to know if the card should behaves as the selected card or no.

- Fix: When deleting a credit card, `onDeleteCard` function
sets `selectedCardNumber` to `undefined` in all cases,
So Make it sets `selectedCardNumber` to `undefined` only if
the deleted card is the selected card.

- Remove some unnecessary codes from `AddCreditCardForm`
componet and make it sets the added card as checkout payment
method by default.

- Cancel `isTheChoosed` option from `deleteCreditCard` function
in `usePaymentMethodsActions`