refactor: rename a hook, states and props related to product details page
Rename `useAreUserBoughtTheProductBefore` hook to `useDidUserBuyTheProduct`,
And rename the request url path "are-user-bought-the-product-before" inside it to
"did-user-buy-the-product", And rename its states variables
`areUserBoughtTheProductBefore`, `areUserBoughtTheProductBeforeLoading`
and `areUserBoughtTheProductBeforeError` to `didUserBuyTheProduct`,
`didUserBuyTheProductLoading` and `didUserBuyTheProductError`.

Rename the prop `areUserCanLetRating` of `ProductRatingSection` component
to `canUserLetRating`.

Rename the props `areUserCanComment`, `areUserCanCommentLoading` and
`areUserCanCommentError` of `ProductCommentsSection` component to
`canUserComment`, `canUserCommentLoading` and `canUserCommentError`.